scrapy/requirements.txt
scrapy
pymongo
python-dotenv
pandas
jupyter
streamlit
plotly
textblob
